Fix bug where parameters were not passed to OLLMod

+++ algorithms/Alg_LinearSU_Mod.h
@@ -55,7 +55,7 @@ class LinearSUMod : public MaxSAT {
 public:
   LinearSUMod(int verb = _VERBOSITY_MINIMAL_, bool bmo = true,
            int enc = _CARD_MTOTALIZER_, int pb = _PB_SWC_,
-           ClusterAlg ca = ClusterAlg::_DIVISIVE_, Statistics cs = Statistics::_MEAN_, unsigned num_clusters = 1)
+           ClusterAlg ca = ClusterAlg::_DIVISIVE_, Statistics cs = Statistics::_MEAN_, uint64_t num_clusters = 1)
       : solver(NULL), is_bmo(false) {
     pb_encoding = pb;
     verbosity = verb;
